+++ src/main.rs
@@ -4,6 +4,7 @@ mod config;
 mod doctor_system;
 mod firewall;
 mod groups;
+mod gui;
 mod logger;
 mod network;
 mod profiles;
@@ -13,7 +14,7 @@ mod windows;
 
 use std::env;
 
-pub const VERSION: &str = "v2.2.0";
+pub const VERSION: &str = "v3.0.0";
 
 fn main() {
     let args: Vec<String> = env::args().collect();